requirements.txt
python-doctr[pytorch]==0.5.1
google-cloud-storage==2.5.0
torch==1.12.0
torchvision==0.13.0
pypdfium2==1.11.0
